Generic css list, update icon
Beginning framework for having generic lists inform the design. nolist is the prototype and everything is commented out b/c it's not working, but it's all set up in the manifest.

faecebook and amazon are sitting files right now.

adding ad removal in the amazon file but it's commented out b/c the parentnode is hard to determine on diferent pattterns.

Updated chrome extension icon to the new icon-48 with soft edges.